Обработка исключений и Spring REST

+++ src/main/java/com/example/job4j_chat/controller/MessageController.java
@@ -25,7 +25,7 @@ public class MessageController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Message> findById(@PathVariable int id) {
+    public ResponseEntity<Message> findById(@PathVariable Integer id) {
         var message = messageRepository.findById(id);
         return new ResponseEntity<>(message.orElse(new Message()),
                 message.isPresent() ? HttpStatus.OK : HttpStatus.NOT_FOUND);
@@ -38,12 +38,18 @@ public class MessageController {
 
     @PutMapping("/")
     public ResponseEntity<Void> update(@RequestBody Message message) {
+        if (message.getId() == null) {
+            throw new NullPointerException("Id cannot be null");
+        }
         messageRepository.save(message);
         return ResponseEntity.ok().build();
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> delete(@PathVariable int id) {
+    public ResponseEntity<Void> delete(@PathVariable Integer id) {
+        if (id == null) {
+            throw new NullPointerException("Id cannot be null");
+        }
         messageRepository.deleteById(id);
         return ResponseEntity.ok().build();
     }
